pyproject.toml
[tool.poetry]
name = "case_extraction"
version = "0.1.0"
description = ""
authors = ["bradley-butcher <bradleydpbutcher@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.62.3"
"pdfminer.six" = "^20211012"
tabulate = "^0.8.9"
spacy = "^3.1.4"
pandas = ">=1.3"
PyYAML = "^6.0"
transformers = "^4.12.3"
torch = "^1.10.0"
sentence-transformers = "^2.1.0"
datasets = "^1.15.1"
prefect = ">=2.0.0a"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
